Move chain verification attribute to self-sovereignty group.
Balances the groups better, matches upcoming presentation contents.

With apologies to @ darrylyeo in #608 having to de-conflict this...
