+++ src/schema/attributes/self-sovereignty/chain-verification.ts
