+++ src/schema/stages/software-wallet-stages.ts
@@ -8,9 +8,9 @@ import { transactionBatching } from '../attributes/ecosystem/transaction-batchin
 import { addressCorrelation } from '../attributes/privacy/address-correlation'
 import { multiAddressCorrelation } from '../attributes/privacy/multi-address-correlation'
 import { privateTransfers } from '../attributes/privacy/private-transfers'
-import { chainVerification } from '../attributes/security/chain-verification'
 import { securityAudits } from '../attributes/security/security-audits'
 import { accountPortability } from '../attributes/self-sovereignty/account-portability'
+import { chainVerification } from '../attributes/self-sovereignty/chain-verification'
 import { transactionInclusion } from '../attributes/self-sovereignty/transaction-inclusion'
 import { feeTransparency } from '../attributes/transparency/fee-transparency'
 import { funding } from '../attributes/transparency/funding'
